requirements.txt
mcp>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
flask>=2.0.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.12.0